Condition CRUD responses
In the cases in which an specific user is involved, could be the situation that that user does not exist. Despite the query execution wil not fail, is necessary make validations in order to don't return a false response from the HTTP request. In those cases is necessary to return an error response to indicate that the operation is executed over an unexistent user.

get specific user: will be validated with the result array size
update user and delete user: will be validated with the affectedRows property from result.